Create CreateAdminAccountAndCyberArk
Creates a separate admin account for a user then creates a cyberark safe, adds the user to that safe, then adds the adm account and cycles the password. There were more for the switches but I didn't feel like sanitizing them so you get the idea. 

Changed around some of the naming conventions as well.